Copyedit "III. Sanctus"
Since the "div." marker is needed again with the same tweak as in the
Introït et Kyrie, move that tweak to the defintion of \div.

Implement switching to Basses I only for the first 42 measures.

Add some cautionary natural signs.

Switch to hyphenation "San -- ctus".

Get rid of one textual "dim."